+++ src/Entity.cpp
@@ -138,7 +138,7 @@ void Entity::loadAnimations(std::string filename) {
 			if (isInt(parser.val)) {
 				int ms_per_frame = atoi(parser.val.c_str());
 				
-				duration = round((float)ms_per_frame / (1000.0 / (float)FRAMES_PER_SEC));
+				duration = (int)round((float)ms_per_frame / (1000.0 / (float)FRAMES_PER_SEC));
 
 				// adjust duration according to the entity's animation speed
 				duration = (duration * 100) / stats.animationSpeed;
